# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'ASE API'
  description: 'This my API to manage Post on a new web site'
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/actualites/{id}':
    get:
      summary: show
      operationId: show
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: 'Data Post !'
                  data: []
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Data Post !'
                  data:
                    type: array
                    example: []
      tags:
        - Endpoints
      security: []
    put:
      summary: update
      operationId: update
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: 'Must be an image. Must not be greater than 2048 kilobytes.'
                title:
                  type: string
                  description: ''
                  example: facilis
                type:
                  type: string
                  description: ''
                  example: repudiandae
                content:
                  type: string
                  description: ''
                  example: sit
              required:
                - title
                - type
                - content
      security: []
    delete:
      summary: destroy
      operationId: destroy
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the actualite.'
        example: 1
        required: true
        schema:
          type: integer
...